% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegThalamus-package.R
\docType{package}
\name{ggsegThalamus-package}
\alias{ggsegThalamus}
\alias{ggsegThalamus-package}
\title{ggsegThalamus: Thalamic Nuclei Atlases for the 'ggsegverse' Ecosystem}
\description{
Thalamic nuclei parcellations for the 'ggseg' ecosystem: the diffusion-based grouping of Najdenovska et al. (2018) and the structural-MRI THOMAS segmentation of Su et al. (2019). Provides 'ggseg_atlas' objects with 2D polygon geometry and 3D mesh data, for use with 'ggseg' and 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegThalamus}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegThalamus/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})
  \item Didac Vidal-Pineiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

Other contributors:
\itemize{
  \item Center for Lifespan Changes in Brain and Cognition (LCBC), University of Oslo [copyright holder]
}

}
\keyword{internal}
